Refactor: Move ObservableArray.ObservableListMapper one level up. Make it package scope. Remove ObservableSet.ObservableListMapper and use ObservableListMapper instead.
