+++ common/src/main/java/bisq/common/observable/ObservableSet.java
@@ -17,68 +17,14 @@
 
 package bisq.common.observable;
 
-import lombok.EqualsAndHashCode;
-import lombok.ToString;
-
 import java.util.Collection;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.CopyOnWriteArraySet;
 import java.util.function.Consumer;
 import java.util.function.Function;
-import java.util.stream.Collectors;
 
 public class ObservableSet<T> extends CopyOnWriteArraySet<T> {
-    @EqualsAndHashCode
-    @ToString
-    private static final class ObservableListMapper<M, L> implements Observer<M> {
-        private final Collection<L> collection;
-        private final Function<M, L> mapFunction;
-        private final Consumer<Runnable> executor;
-
-        private ObservableListMapper(Collection<L> collection,
-                                     Function<M, L> mapFunction,
-                                     Consumer<Runnable> executor) {
-            this.collection = collection;
-            this.mapFunction = mapFunction;
-            this.executor = executor;
-        }
-
-        public void add(M element) {
-            executor.accept(() -> {
-                L item = mapFunction.apply(element);
-                if (!collection.contains(item)) {
-                    collection.add(item);
-                }
-            });
-        }
-
-        public void addAll(Collection<? extends M> values) {
-            executor.accept(() -> values.forEach(element -> {
-                L item = mapFunction.apply(element);
-                if (!collection.contains(item)) {
-                    collection.add(item);
-                }
-            }));
-        }
-
-        public void remove(Object element) {
-            //noinspection unchecked
-            executor.accept(() -> collection.remove(mapFunction.apply((M) element)));
-        }
-
-        public void removeAll(Collection<?> values) {
-            //noinspection unchecked
-            executor.accept(() -> collection.removeAll(values.stream()
-                    .map(element -> mapFunction.apply((M) element))
-                    .collect(Collectors.toSet())));
-        }
-
-        public void clear() {
-            executor.accept(collection::clear);
-        }
-    }
-
     // Must be a list, not a set as otherwise if 2 instances of the same component is using it, one would get replaced.
     private transient List<Observer<T>> observableListMappers = new CopyOnWriteArrayList<>();
